+++ CHANGELOG.md
@@ -10,11 +10,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Create 2024.4.0-RC.2 build
 - REMIX-2593: Added a centralized TreeWidget with additional logic
+- REMIX-3075: Added a layer type check in validation
 
 ### Changed
 - REMIX-2593: Changed the LayerTree widget to work with multiselect
 
 ### Fixed
+- REMIX-3076: Added trailing slash to end of dirname when double-clicking in file dialog
 
 ### Removed